Second commit, create/modify/GlobalErrorHandler

+++ src/main/java/pet/store/entity/PetStore.java
@@ -21,7 +21,7 @@ import lombok.ToString;
 public class PetStore {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int petStoreId;
+	private Long petStoreId;
 	
 	private String petStoreName;
 	private String petStoreAddress;
